json_object_merge: Don't create deep copies.
Why should we when there's a reference counter? Sure, you might want to re-use and change the original object... but in every single instance this function is currently used, the new array is trashed immediately after merging, anyway.

Also, change the function parameters to make a bit more sense, and remove that part about array merging - this has become such an integral part of the whole spec by now.