Add json_array_from_wchar_array().
Not too happy about either naming or placement... but well, that code appeared in both the loader and the configuration tool, so it deserved to get its own function.

Putting it into win32_utf8, returning a char**, would have been the alternative, but that has to be recursively freed, so I decided against it. GetCommandLine() and CommandLineToArgvW() are basically the same.