Load dialog resource replacements from external binary files.
When translating resources using a tool like Resource Hacker, the location of a resource inside the executable is very likely to change. That makes the initial plan of using binary hacks for complete modifications of the resource script completely unreliable.

However, this introduces a much nicer system: Now, we can just directly supply that blob as a separate file, using the same file name scheme as for the JSON dialog descriptors. Elegant, native, understandable, and with just a few lines of additional code to pull it off.