Advanced image patching, part 2: Alpha-based patching on sprite level.
By classifying an alpha channel into one of three states ("empty" if all alpha values are 0, "opaque" if all alpha values are 0xff, and "full" for everything else), we can selectively patch parts of images, defined by the ANM entry's sprite list.

This can drastically reduce the size of the replacement PNGs in certain cases: If only certain sprites of an image need to be edited, the unchanged ones can be "ignored" by simply deleting them (that is, zeroing their alpha channel). Also, it paves the way for PNG stacking.

The alpha analysis is implemented in a rather naive way, though...