Introduce a system to allow customization of files loaded during resolving.
Previously, the sequence of generic and build-specific files had to be hardcoded into every single resolving function. This new "resolving chain" system keeps this sequence of file names in an array.

Not only does this eliminate a great deal of redundancy when writing new file resolvers for new formats. This also allows adding new files to the chain to customize existing resolvers for certain file formats, which would previously have required duplication of code.