Move thcrap updating information from patch.js to the run configuration.
In a time without both global.js and the slightest idea of how popular the engine (not the patches) will eventually be, patch.js was of course the most obvious place to put that. However, now that we've actually *come* across a situation where there will be multiple download URLs, and it turns out that the best way to handle this is to just let patch stacking take care it.

Also choosing better key names.