sprite_patch_set(): Verify the destination sprite coordinates.
Fixes a bug introduced with advanced image patching in 2014-01-03. ANM entries can define sprites with coordinates outside the dimensions of the original texture. However, if these happen to be inside the dimensions of a replacement image, the sprite is "patched" regardless - thus, overwriting random memory in the process and crashing the application.

This broke the trial versions of th14 if thpatch's Western name order patch is used.