Revise detour chaining to use a system similar to MS-DOS interrupt chaining.
It's just another normal day of development. Today, a new set of detours are on the agenda. Everything is working fine, but there is that one instance of duplicated code you'd like to remove, which gets you thinking. "Hm, detour_next() is nice and all, but wouldn't it be cool if we could somehow get a pointer to the next function in a chain...?"

Then, you have an epiphany.

"Wait a moment... Why didn't I just copy the design used for interrupt chaining in MS-DOS?"

And indeed - after rewriting the whole system, everything just feels so much nicer. And you got your pointers.

Wow.

----

So, by reducing these chains to an implicit sequence of function pointers kept separately by every module, we remove the need for a function like detour_next(), fixing *all* the drawbacks of detour chaining in the process:

* No more platform-dependent assembly to explicitly push a variable argument list to functions with a constant parameter list.
* The function pointers now used to continue the chain can be explicitly initialized with a fallback function. Since these are resolved at module initialization, they are therefore guaranteed to work.
* Last but not least, this also gets rid of the hashtable and array lookups on every invocation.