Only download necessary patch data during configuration and updates.
And that would be all of fast bootstrapping (#36) we can reasonably implement
without a GUI.

... except that it's not yet as fast as it could be, due to a slight flaw in
the current patch file structure: "<game>.js" and "<game>.<build>.js" are
stored in the root directory of a patch. The patches themselves don't include
a simple list of the games they support (and really, why should they), so
determining the global and game-specific parts of a patch becomes slightly
less trivial, as there is no way around building such a list internally by
parsing through files.js.

If we are about to break backwards compatibility anyway, we should move these
files into the subdirectories of the respective games (so that e.g.
"th14.v1.00b.js" becomes "th14/th14.v1.00b.js") for even faster bootstrapping.
Then, global files are easily identified by the lack of a slash in their
filename.

This implementation already pretends that this is the case. In a typical setup
where only a single game is to be patched, this would still download 40-ish
JSON files too many.