Fix a few small syntax issues that throw errors or warnings in GCC.
For now, this is everything we can do that doesn't break the Visual Studio
build in turn. I'd like to be able to actually *test* the build before
committing a solution for for how we're going to handle breakpoint_process()
in GCC, and since I've pretty much given up on porting this project to CMake
or qmake myself (see #39), that's not going to happen in the foreseeable
future.