thcrap_loader: Give "exe" precedence over "game" in the run configuration.
Not that it matters much, but games.js is a rather central file and run
configurations are more specific, so this order makes more sense to me.

inb4 'how about merging "game" and "exe" into a single value, don't the
command-line arguments kind of behave the same': Remember that the thcrap core
uses that exact "game" value to construct the file paths for game-specific
data, and the run configuration is always merged on top of the final
configuration built from global.js, $(game).js and $(game).$(build).js.