thcrap_loader: Go back to using Jansson instead of the QJson* types.
Still not sure which one of the two JSON libraries to use for the GUI code.
The differences in performance and line count are pretty much negligible.
Therefore, it all comes down to whether we either want to duplicate all the
patch stacking logic by writing a second version of it using the QJson* types,
or put the effort into cleanly separating the platform-specific parts from the
existing code in the thcrap core and adding Qt and Win32 backends.

Getting completely rid of Jansson and rewriting all existing code in QJson*
is a pretty bad option, though: Since the Qt GUI code is basically cross-
platform, there's nothing in the way of having native Linux and Mac OS builds
of those parts of thcrap, which would come with a nice native GUI in place of
an ugly Wine emulation. If the Win32-specific core of thcrap continues to have
no dependencies on Qt, Linux packages then wouldn't even need to include the
Qt 5 Windows DLLs that are part of the Windows distribution. Instead, their
packages would merely consist of thcrap, win32_utf8, Jansson, libpng and zlib
as Windows DLLs, thcrap_configure, thcrap_loader and thcrap_update as native
Linux applications, and dependencies on Qt 5 and Wine (as well as potentially
Jansson).