inject: Don't pass an invalid current directory to CreateProcess().
This is the sort of thing that maybe should have better been fixed directly in
win32_utf8. But well, CreateProcess() is supposed to return a distinct error
code for exactly that case, so it might not be the best idea to silently do the
right thing. (Which may not be the right thing at all for certain applications.)