inject: Only apply detours for newly loaded DLLs a single time.
Mainly necessary to keep certain applications that LoadLibrary() the same DLL
over and over from bloating the log file (and the on-screen console, if
activated).

Also leads to a drastic performance increase, even if the console is disabled.
In my main test case, TH09 with the DirectX 8 → DirectX 9 converter from

	http://enbdev.com/convertor_dx8_dx9_v0036.htm

this reduces the time spent in our LoadLibrary() hook after a full, Skipgame'd
run from 932 ms down to 11 ms. (Which I really didn't take for granted, given
that Jansson still does a malloc() and a free() when updating a value in its
hash table implementation.)