Rewrite the breakpoint prolog and epilog code in Assembly.
Splitting breakpoint prolog/epilog and dispatch into two functions (one in
Assembly, one in C) actually ends up much cleaner than having both in a single
function with __declspec(naked), which GCC doesn't support for x86.

For the few lines of assembly in thcrap, we don't really need an assembler
with lots of features. Therefore, just using GAS, which is bundled with MinGW
as part of binutils, is the best choice here. It even has a .INTEL_SYNTAX
directive. ☺

For now, I'm also including a i686-w64-mingw32-as.exe and the necessary
MSBuild customization files (adapted from VSYASM) to keep everything
compilable in Visual Studio during the transition to GCC.