Make bp_entry.asm SAFESEH-compliant.
Strangely, thcrap did link without this for a while, but then it didn't
anymore, and it breaks for brliron too. This fixes the problem the hard
way, because it felt like quite a useful thing to be figured out.

So again, for Google, since this took way too long to find out: In order
for an object file compiled with MinGW's GAS to work with the /SAFESEH
option of Visual Studio, all you have to do is to define a non-global
symbol named @feat.00 as 1, like this:

	.set @feat.00, 1