Always set file_size to 0 for stack file resolution in empty stacks.
Fixes a crash with the combination of thcrap_tasofro and empty patch
stacks.

thcrap_tasofro doesn't initialize its file_size variable when resolving
fileslist.txt, so it runs LoadFileNameListFromMemory() with a NULL
pointer and an uninitialized size value. Setting file_size to 0 on
failure is definitely something we should have done all along, so that's
how we're going to fix this.