Update our dependent libraries.
Including an important change to Jansson that (finally) preserves the
insertion order of JSON object items, which means that we can now
better diff log files, get a consistent repository order in
thcrap_configure, and probably many other small things I can't think of
right now.

However, we're only updating libpng up to 1.6.25, as 1.6.26 introduces
an incompatibility with zlib-ng: Turns out that the original zlib
actually *is* still being developed, just not on the master branch.
There, it recently introduced a function named inflateValidate(), and
libpng 1.6.26 uses thes this function if zlib identifies itself as
version ≥ 0x1821. zlib-ng is version 0x128f, but doesn't come with
inflateValidate(), so libpng doesn't compile. I'll ask upstream about
this, but given that zlib-ng hasn't seen any development in the last
few months, I guess we'll be switching back to the regular zlib soon.