Implement log file rotation.
Thanks to a recent change to Jansson that preserves the insertion order
of JSON objects, diffing thcrap's log files is now actually useful, so
we might as well keep more of them around by default. The only remaining
inconsistency of patch updates being printed randomly, due to them
running in a separate thread, will be fixed later with #37.