TSA: ANM: Clamp width and height of sprites to the actual texture. [V]
So not only can ZUN specify the sprite X and Y coordinates to be outside
of the actual texture, but he can do this with width and height as well.
This fixes the obvious game crash that happens when trying to patch such
a sprite. Thanks to @DTM9025 for reporting this issue!