BP_file_size(): Remove the conditions around the file size assignment.
th09 needs the POST_JSON_SIZE to be unconditionally written out to a
scratch register inside its file decryption function, where the final
decrypted buffer is allocated. The actually intended size register is
used as the loop counter for the decryption, and putting anything else
there will result in a few bytes of corruption at the end.

So this allows us to place two additional file_size breakpoints to work
around that fact, once the next build is out.