zip: Iterate over extra fields correctly. [V]
Instead, it was my embarrassing 2014 insistence on avoiding any
allocations in this code, no matter how small, even if that meant doing
two things at once and needlessly cluttering the code.

Fixes the issue of the 2017-04-01 build not extracting correctly during
a self-update, which showed a quite misleading error message.