TSA: MSG: Move the encraption functions into th06_msg.c.
So, rather than depending on thcrap's specs module, and thcrap's
jsondata module, and thcrap's jansson_ex.c, which by definition has to
include GetModuleHandle(), and then setting up the build for both thcrap
and the scripts… yeah, let's get better get rid of formats.js and
hardcode these instead. Especially since it rarely changes these days,
if at all. But if it does, it'll probably be a significant change that
requires a code update anyway.

And it'll be way easier to debug.

(Really, formats.js stopped making sense as soon as we had automatic
self-updates.)