textdisp: Add font quality as a fifth parameter to font rules. [V]
Which effectively saves us… a grand total of 1 binary hack per
game, because it only works as intended with the ruby font. Dialogue is
unfortunately printed with the same font as spell cards, endings, and
the Music Room, which we *don't* want to reduce the quality of. So not
only would this need one binary hack for changing the in-game dialogue
font, we'd also need *another* one to differentiate the new dialogue
font, so that thcrap can apply the rules… which are then pointless
because you might as well just include them.

And in the end, the text still looks like shit.

Oh well, ruby legibility *is* kinda important, so let's still go with
this.

(Or maybe we can inject a new font switch command into each dialogue MSG
file. Then we'd save that one hack. Whatever, just wait until I actually
get to that feature, will you? Thanks.)