update: Reword the old version message.
Because "confirmed to work" signals perfect support.

The *actual* truth is that we support the old versions of every ZUN game
released since thcrap has been a thing (that is, May 2013), but not of
every Tasofro game… which is way too complicated to mention it in a
message box that should just get people to update their games.