Add a separate mechanism to disable hackpoints, to allow easy opt-in.
Opting *out* of a hackpoint is quite straightforward, you simply null
out any of the vital parameters – the address, the code (for binary
hacks) or the cave size (for breakpoints). Opting *in*, however, would
then mean that the providing patch leaves one of its vital parameters
null and expects the patches that opt in to fill out the correct,
agreed-upon value themselves. But what if the providing patch wants to
later change the hack?

So, better have a separate mechanism to make both opt-in and opt-out
more explicit.

(The specific use case: Some languages prefer to flip the two lines of
the in-game "new nickname" pop-up in TH14.3. We want to provide this
(initially disabled) hack in base_tsa so that any language can make use
of it.)