Standardize the distinction of pointer and value parameters in breakpoints.
The main benefit of json_immediate_value() (the former
json_register_value() from thcrap_tsa's gentext module) is that it
allows you to "hardcode" a certain value in the parameters of a
breakpoint, rather than being forced to pull it from a register. This is
something that all breakpoints can benefit from, and it might eliminate
the need for some of the "preparation" binary hacks.

However, turns out that also we need an *actual* json_register_value()
for the upcoming support of multiple output pointers in gentext
breakpoints, using a flexible array. So, let's extend it from registers
to pointers in general, while we're at it.

This naming scheme will also work nicely once we support expressions
such as "esp+8". Even though "immediate" is certainly not the best name,
but I spent hours and couldn't come up with anything better…