xpcompat: Add a wrapper for thread-local structures.
Oh look, something that MinGW/GCC would have actually done for us, with
its support of the _Thread_local specifier introduced in C11. Then
again, this is not exactly rocket science, and it's dwarved by all of
the things MinGW/GCC does *not* do for us, most notably its lack of
support for #pragma comment(lib) or DLL delay-loading.

Fixes crashes when thcrap_tsa's gentext breakpoint is hit on Windows XP.
(Which no one reported, which means that we might just be able to stop
supporting it soon.)