inject: Correctly reference-count DLLs by not reference-counting at all. [V]
Haha, seriously, how stupid was I? We don't care about any reference
counts above 0 or 1 to decide whether we want to detour the imports or
not. And for those counts, there's GetModuleHandle()… which also takes
the initial state of imported DLLs into account.

This fixes the crash during the loading screen in TH09.5 when it is
started in fullscreen mode. This is roughly what previously went on
there (I didn't investigate this fully after I found the cause):
• Direct3D 8 uses ANSI internally.
• ZUN calls CreateWindowExA(), which is wrapped by win32_utf8.
• ZUN calls IDirect3D8::CreateDevice(). Direct3D 8 subclasses the window
  proc using the unwrapped GetWindowLongA(), SetWindowLongA() and
  CallWindowProcA().
• The game proceeds to the loading screen.
• ZUN calls LoadLibrary("d3d8.dll"), in order to get the address of
  DebugSetMute(). (Yeah, GetModuleHandle() would have sufficed there,
  but that's beside the point.) thcrap's LoadLibrary() detour kicks in
  and applies the detour list to all imports of d3d8.dll, including the
  USER32.DLL functions.
• Direct3D 8 suddenly calls GetWindowLongW(), SetWindowLongW() and
  CallWindowProcW() instead.
• ???
• The game crashes.

This also allows us to procrastinate the choice whether we actually want
to recursively detour all imported DLLs on startup. (And all the
disasters that probably come with that.)

By comparing the previous module handle to the newly loaded one, rather
than merely comparing the old one against the nullptr, this also seems
to observe one of the caveats with GetModuleHandle():

	"If `lpModuleName` does not include a path and there is more
	than one loaded module with the same base name and extension,
	you cannot predict which module handle will be returned."

Even in that case, both GetModuleHandle() and LoadLibrary() seem to
always return the same module given the same path.