devicelost: fix using more generic approach
Apparently my original fix only worked for me, because there are
multiple sets of Reset calls, and I've only patched one of them.

To avoid going through all the executables again, I decided to just
detour the Reset function. Idealy we would need to do something like
nmlgc/ProxyDLLs, but this should work for now.

I tried copying the vtable, but it turned out to be a bad idea, because
the size of vtable is unknown. The only other way, besides creating a
proxy object, is editing the vtable directly.