Breakpoints: Rewrite to run in O(1) and to not use global state.
Our n is still way to small for this to actually make a difference. With
TH14.5, the heaviest user of breakpoints so far, this only reduced the
time spent in breakpoints by 2.5%. In TH06, this was even 12% slower?!

Then again, that probably was old and broken support code. Turns out
we'd like to have parts of it for DRM cooperation, as we need two
distinct sets of breakpoints active at the same time.