TSA: Shift speech bubbles to make room for longer lines, if necessary. [V]
This always was somewhat of an issue, but with the two bosses in TH16's
Stage 5, it really became bad for a number of languages.

However, the tip will still remain at the end of the bubble, and it's
impractical to hack around that. The longer the text, the more the tip
will therefore be shifted away from the portrait of the speaking
character. Make the text long enough, and it might misleadingly point to
that other character standing right of the speaker, or even to the
interface.

So it should only be used sparingly.