TSA: MSG: Support TH14's new speech bubble shape/direction command. [V]
As of TH16, this is only used in the TH14 Extra Stage for the pre-
midboss dialog, to make Benben's one bubble extend to the right, despite
it being set up with opcode 8 before.

Without this, the MSG patcher thought this was a standard right-side
bubble extending to the left, and since it is positioned quite far left
at 112 pixels, it erroneously shifted the bubble quite a bit to the
right for most translations.

This new opcode could also change the shape of the bubble, but for some
reason, the text for all left-side bubbles is weirdly "indented" by
about 40/80 pixels. Not our problem though, and it's unused anyway.