update: Send the operating system version in the User-Agent. [V]
Turns out we *really* want to stop support for Windows XP now. We'd like
to switch the patch data servers to HTTPS, but wininet on XP couldn't
use those anymore unless we enable outdated protocols and ciphers that
really shouldn't be used anymore.

So, let's collect a bit of data and see whether we could actually stop
supporting XP.