thcrap: Add Steam integration for games without. [V]
Because y'know, ZUN might have been lazy enough to not link to that DLL
and call a single function.
Or maybe he wasn't, and all of this was for nothing, because Tasofro
certainly won't be.

And yes, this means that thcrap now comes with a proprietary blob by
Valve Corporation. But before all the GNU/people (and Pooka) freak out:
• `steam_api.dll` will only ever be loaded if Steam is already
  installed. In that case, you already agree to running binaries by
  Valve on your system.
• Just as with `thcrap_update.dll`, you can always delete
  `steam_api.dll` to deactivate this feature… except that you have to
  re-delete it for every new build. Maybe we'll fix that too, let's see.