TSA: MSG: Correctly align TH13's Parallel Ending explanation. [V]
Since TH13 repurposes opcode 9 for this type of "ZUN message", centering
would have been really nice… except that the center of the playing field
is *not* the center of the screen, and we're therefore robbing the
translators of 192 pixels of screen space.
Sure, we could have only centered the text if it was short enough, but
then we would have had to divide the padding by two just in this case,
and the sprite isn't animated to expand from the center anyway… so let's
just keep it simple there.