thcrap: Support custom base addresses for the hackpoints of init stages. [V]
Will eventually fix compatibility with Kaldaien's SpecialK mod,
SteamStub DRM, and thcrap. Not only is 0x10000000 the default base
address of every DLL compiled with Visual Studio, SteamStub is also
perfectly capable of relocating this DLL if necessary… like, for
example, if some proxy DLL already claimed that base address for itself.

It may look weird to have a breakpoint from a previous stage temporarily
change the base address for the following stage, which is necessary in
case it needs to be a VirtualAlloc()'d DLL that can't be scoped to by
name. But well, you can always add a comment JSON string to explain
that.