update: Ignore servers without an URI scheme.
Might have been nicer to check the errors returned from
InternetCombineUrl() or InternetCrackUrl() instead, if only they were
reliable.
Having a "default protocol" doesn't seem too useful anyway, so it's
better to just ignore the server on object creation already. Which is
also where we might want to split off the protocol part in the future,
to control things like HTTPS-to-HTTP fallback, for example.