thcrap: Log: Show an error message if thcrap_log.txt can't be written to. [V]
The challenge here was avoiding a second message box if the user decides
to continue and uses vpatch. As it turns out though, this is once again
the best kind of reference counting where we only care about 0 or ≥1, so
we only need to *create* a 1-byte file mapping and not actually write a
reference counter to it.