update: Move wininet initialization and cleanup out of DllMain. [V]
Haha, no, e17a496b48983bf02ca219c81b5fe7e286a7c98f, *this* was the
reason! Fixes rare hangs when starting thcrap_configure.

Now, the interesting question about this was what to do with
InternetCloseHandle(), which isn't safe to be called from DllMain()
either. Do we even need to call this for the handle returned from
InternetOpen()? Sure, process shutdown is supposed to be "the ultimate
cleanup", but how can we be sure that wininet doesn't keep this handle
around in a process-independent manner? Might be overly paranoid to say
this, but hey, I've just fixed a rare bug that only was reported after 8
months, so no, we can't.

So we need a manually called, exported shutdown function, plus a
_mod_exit function for while this is still a plug-in. And while we're at
it, let's standardize the former as thcrap_update_exit() right away, to
be prepared for supporting more protocols in the future.

(Having only a single thcrap_update_mod_exit() probably looks weird if
people have to call a supposedly automatically called hook function
manually.)

… Oh well, but if we Ctrl-C or X out of thcrap_configure, it won't be
called anyway?! I way overthought this once again.

Fuck this, I'm never going to use wininet ever again if I can help it.