configure: Extend Y/N questions to accept multiple and customizable answers.
Hey look, a use case for std::array! Too bad it still means that the
function has to be a template, and that it must be defined in a header
file, or you get linker errors.

I would have also preferred a plain char array, but initializer-list
doesn't convert to it.