update: Redownload checksum-less invalid JSON files from different servers. [V]
Previously, there was only a single chance for a files.js download to
return a HTTP 200 OK response, and if that one was invalid JSON, that
was it, and the patch update was immediately aborted.

The new comparatively "silent" way of just reporting an error to the log
file also provides a better user experience in case people are using
proxies that intercept HTTP (as in, non-HTTPS) traffic – no more
misleading message boxes about invalid JSON in this case, just an
automatic fallback.

Also applied it for repo.js while I was at it, even though it isn't
really useful yet, since repositories are currently only discovered
using a single URL. That might change in the future, though.