TSA: Cache fonts for TH06-TH09 rather than re-creating them for every line. [V]
Not only is this necessary to fix misalignment after font changes and at
the beginning of the game, it also removes the potential framerate drop
that came from re-creating and destroying the font for every line of
text rendered. Accumulated over a full 6-stage run, the stock games
wasted:

• TH06, base_tsa, MS Gothic: 1359 ms
• TH06,  lang_de,   Aroania: 1080 ms
• TH07, base_tsa, MS Gothic: 1651 ms
• TH07,  lang_de,   Aroania: 1603 ms

Funded by zorg.