thcrap: Remove the file access critical section.
This was meant as a poor attempt to eliminate race conditions between
the updater and everything else. If a load and write to the same file
collided and the timing was right, the load *might* have blocked until
the write was done, and directly loaded the new file afterwards…

…well, back when the updater still ran in the same process. Now, this
doesn't really make sense anymore, and BGM modding needs streaming
access to more than one open file handle at a time. If we want something
like this again in the future, it should also work with any other
process accessing a file.

Funded by Splashman.