TSA: ANM: Correctly wrap internal sprites that start outside the texture. [V]
So yeah, I fixed that logic error of vertical sprites larger than the
texture height not being split correctly in 84470af… but apparently
never considered that sprites can also *start* at positions larger than
the size of the texture, in any dimension? Which would always need to be
wrapped for our purposes? And we never did? Which ended up breaking
sprite-level patching of front/ename.png in every game since TH14?

Thanks to BraviNoname#9568 for reporting this bug.