BGM modding: Implement a custom file resolver. [V]
By turning the codec hierarchy into a resolve chain, we not only
establish it within a patch, but allow BGM mods to be stacked
independently of the codec chosen. So, an MP3 from a higher patch would
always override a FLAC from a lower one. Also, this forces the loop file
to use the same codec as the intro file.

Funded by Splashman.