BGM modding: Add MP3 support. [V]
So MP3 BGM actually is no problem and can loop as well as the other
formats… *if* you cut the loop correctly in the first place.

Error messages on invalid files will be rather suboptimal, because MP3
apparently is so loosely specified that pretty much anything can be an
MP3 file, and confirming the opposite is equivalent to solving the
halting problem?
https://sourceforge.net/p/mpg123/mailman/message/34642389/

Funded by Splashman.