TSA: Perform a bunch of intricate ASCII string alignment hacks for TH06. [V]
Because even I got tired of doing all these things in ASM, and then
doing them again for the trial version. Just look at all this stuff!
That might look like a lot of heavy-handed boilerplate, but we do want
to support more than just TH06 eventually, and we're dealing with
binary hacks that turn parts of the same function into a callback for
us… better to front-load as much of the pointers as possible there to
keep the binary hacks as reusable as possible.

Anyway, that (more than) concludes The Virgin Bare Minimum EoSD
Support™! I can now confidently recommend thcrap for TH06 over static
patches.

Funded by zorg.