thcrap_i18n: lib for translating thcrap itself
Note to bruno: Make your scripts bundle thcrap_i18n.dll. Also bundle
i18n directory in the same directory as the dll, but that's not urgent
(the library should be no-op when that directory is missing)

Currently the only way to change language is to delete i18n/config.json
I'll create an extra button for invoking the language change dialogue,
but keep in mind that this will require application restart.

Adding support for new languages is done in plurals.cpp.

Funded by rosenrose.