Move win32_utf8 to libs/ and write our own vcxproj for it.
win32_utf8 have been designed to be usable by other projects. I think putting it with the other libs better reflects that.
And, for the same reason, we use a copy of the .vcxproj, as with the other modules. We have some requirements for the thcrap build that doesn't fit a general-purpose library, like building with an outdated compiler (Visual Studio 2013 in XP compatible mode) or using bin/binaries as the output directory (we don't do that yet, but we will probably do it soon). Using a copy lets us write anything we want in the copy while keeping the original pure.