thcrap_update: look for the latest version number in thcrap_update.js
Now that different branches and milestones are a thing, the process that triggers an update must change, too.
This was either way easier and way harder than i thought: moving the version check from notify.cpp to self.cpp was pretty easy, but this broke the messages since notify didn't have infos about the build anymore.
Now is self that provides the target version to notify. This allows for more accurate message boxes. Now the user can't mistake an incremental update for an update loop.
There's an unpleasant side effect: if thcrap can't retrieve the build information from the server, it will trigger a blocking message box. On the stable build it can be triggered in some limit (offline) scenarios, but now it happens every time you're offline for any reason.
I'm considering adding an additional configuration that allows to skip error messages related to version-checking.