thcrap_update: self.cpp now writes on log when looking for an update
This has mainly two effects:
- Makes all update procedure look similar in the log file (each repo logs when the process starts)
- Makes clear when thcrap enters self_update

Actually one log line is generated when trying to download "thcrap_update.js". My "problem" with this - and it's not even a real problem tbh - is that this line belongs to the "network" component and not to the updater. It works fine right now, it might not in the future.
I originally planned to log more, but it's not really that necessary.