Fix/hide warnings.
There was a lot of them, and it was bothersome to filter the new, serious ones from the old useless ones.

We know XP support will be removed next year, no need to spam the console with this.
I'm not fixing warnings in other projects.
Most other ones were small type and cast problems.