Specify calling convention for strings_lookup
Because before the calling convention was never specified, the compiler
could always decide to suddendly use __stdcall with that function. This
would cause serious issues with th09 because a base_tsa binhack relies
on that function using the __cdecl calling convention. If suddendly the
calling convention would change to __stdcall, everyone's th09 would fail
to launch because of that binhack. Because of that I specified the
__cdecl calling convention for strings_lookup so that it will always
stay __cdecl