Add thcrap_test
Yes, we now have some automated testing. That's not a lot, but now, when people write new features, they can write tests for them. And I plan to write tests for most of the parts changed by the runconfig commit, or at least the patch and stack functions. Covering some of the configure functions would be good too - we already had some errors in them, and we still have some use after free.