netcode: stop using a server only on server errors
A 503 or 504 error code, a timeout, or a connection refused are
usually a sign that a server is dead, and we don't want to try them
again in this session. On the other hand, a 404 error is fine,
it only means that a file is missing, the server still works -
and, even if it means a patch is dead, other patches on the server
might very well be working. As for when a download is cancelled,
this is a user action and doesn't mean the server is dead.