netcode: report progression to the caller.
And because reporting progression goes through all the layers, with every layer adding it's own bit of information, and because reporting progress is also about reporting completion, it's time to think a bit more about how the layers talk to each other.
Especially for completion, there are 2 choices: either use a return value, or give the return value through a callback. But before that, there's another choice: when and how do we report start, completion and failure to the callback? Reporting completion as a return value for the lower layers can be fine, but the upper layer need to have a callback called every time a file is downloaded. We can either have the lower layer report every status change, or have the lower layer report only progress through a callback, and have the middle layer call the callback before calling the lower layer and after the lower layer returns. But it felt awkward for several reasons:
- The lower layer is a good place to report start and end, it can do that easily (just put a call at the beginning and at the end of the function). Reporting that in a middle layer can be harder.
- We have less state in the middle layer. For example, the Downloader just say "download that file from one of these servers", and the File decides which server it will choose. And the URL is something nice to log, especially on failure.
- We need to return several values. At least the result, and an error code in case of failure. Sure, we can return a tuple, or add the error code in the File class. But it's easier with callbacks.
So it seems better to have the lower layer report completion and failure to the callback directly.
And now that the middle layers have their completion reported in a callback, it felt weird to just forward the status to the upper layer callback and use a return value to look at the completion ourselves. So here we go, in full callback mode.
At least until the end. We still want stack_update and global_update to return only when everything is finished. On the other hand, we don't need them to return anything, we only need them to wait (after all, they are returning void). So we wait until everything is done. Status is still reported through callbacks.

As for the choice of having 3 callbacks instead of one, I started with one and everything was screaming "make it 3 callbacks". The callbacks were always splitted in 3 parts (progress, success, or failure), every part had its own parameters (failure has an error code, progress has a progress status, and success has the downloaded file), so it just made sense.