netcode: remove PatchUpdate class.
It was already acting as a part of the Update class (it even had a
"friend" to access the Update private state), and the only thing
that made it really different from the Update class was...
one variable. We just pass it as a parameter to 2 functions
and we're done, we can remove that class.