netcode: improve repo discovery
A long-running request we have is to *not* write every repo.js in the world to the disk. These files are small so it doesn't really waste disk space, but is creates 15 directories, which gives a feeling of clutter and wasted space, and makes finding a repo in the directories list harder.
It isn't really useful for offline discovery either, because if you're offline, well, that's great, you have that huge list of awesome repos, but you can't download any of them because you're offline...
So now, during the discovery, we keep the repos in the RAM, and we write them only in patch_bootstrap(). That function already has to create the repo directory while creating the patch directory, so it doesn't hurt to write the repo there. And while that function isn't available in offline mode, you won't find any new repo in offline mode, so you won't need to write any.

Having an in-ram list of explored repos also have the benefit of being a cache. We have this list of repos we already explored, we can use it to avoir re-explring them.

Because we now return the list of repos instead of storing it on disk, it became bothering to have 2 separate functions to return repos. And it made me realise that having 2 functions is useless anyway. Merging them have the benefit of sharing the RepoDiscover object, and so sharing the URL cache (which had to be passed as parameter in the nmlgc implementation, and wasn't shared at all in mine).